rename: 커뮤니티 핸들러 경로 수정
