+++ backend/community/serverless.yml
@@ -31,7 +31,7 @@ provider:
 
 functions:
   createPost:
-    handler: service/createPost.handler
+    handler: src/handlers/createPost.handler
     events:
       - http:
           path: /{boardType}
@@ -39,7 +39,7 @@ functions:
           cors: true
 
   getPost:
-    handler: service/getPost.handler
+    handler: src/handlers/getPost.handler
     events:
       - http:
           path: /{boardType}/{postId}
@@ -47,7 +47,7 @@ functions:
           cors: true
 
   listPosts:
-    handler: service/listPosts.handler
+    handler: src/handlers/listPosts.handler
     events:
       - http:
           path: /{boardType}
@@ -55,7 +55,7 @@ functions:
           cors: true
 
   updatePost:
-    handler: service/updatePost.handler
+    handler: src/handlers/updatePost.handler
     events:
       - http:
           path: /{boardType}/{postId}
@@ -63,7 +63,7 @@ functions:
           cors: true
 
   deletePost:
-    handler: service/deletePost.handler
+    handler: src/handlers/deletePost.handler
     events:
       - http:
           path: /{boardType}/{postId}
@@ -71,7 +71,7 @@ functions:
           cors: true
 
   createComment:
-    handler: service/createComment.handler
+    handler: src/handlers/createComment.handler
     events:
       - http:
           path: /{boardType}/{postId}
@@ -79,7 +79,7 @@ functions:
           cors: true
 
   listComments:
-    handler: service/listComments.handler
+    handler: src/handlers/listComments.handler
     events:
       - http:
           path: /{boardType}/{postId}/comments
@@ -87,7 +87,7 @@ functions:
           cors: true
 
   getComment:
-    handler: service/listComments.handler
+    handler: src/handlers/listComments.handler
     events:
       - http:
           path: /{boardType}/{postId}/comments/{commentId}
@@ -95,7 +95,7 @@ functions:
           cors: true
 
   updateComment:
-    handler: service/updateComment.handler
+    handler: src/handlers/updateComment.handler
     events:
       - http:
           path: /{boardType}/{postId}/comments/{commentId}
@@ -103,7 +103,7 @@ functions:
           cors: true
 
   deleteComment:
-    handler: service/deleteComment.handler
+    handler: src/handlers/deleteComment.handler
     events:
       - http:
           path: /{boardType}/{postId}/comments/{commentId}
@@ -111,7 +111,7 @@ functions:
           cors: true
 
   likePost:
-    handler: service/likePost.handler
+    handler: src/handlers/likePost.handler
     events:
       - http:
           path: /{boardType}/{postId}/like
