+++ backend/community/src/handlers/getPost.mjs
@@ -3,7 +3,7 @@ import {
   UpdateItemCommand,
   GetItemCommand,
 } from "@aws-sdk/client-dynamodb";
-import { createResponse } from "../util/responseHelper.mjs";
+import { createResponse } from "../utils/responseHelper.mjs";
 
 const ddbClient = new DynamoDBClient({ region: process.env.REGION });
